+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `instanceId` parameter to `hostCreate` method.
 - Update `smart` parameter to `kind` for inventory methods.
 - Fix `inventoryModify` method by adding mandatory `id` parameter.
+- Add `credentials` and `executionEnv` parameters to `jobTemplateLaunch` method.
 
 **GoSS**
 - Add `logLevel` parameter to `server` and `validate` methods.